fix issue where JsonSyntaxException wasn't exposed
